[organize] make several modules
